rewrite auto show image selector
