+++ src/muya/lib/contentState/index.js
@@ -26,7 +26,6 @@ import tocCtrl from './tocCtrl'
 import emojiCtrl from './emojiCtrl'
 import imageCtrl from './imageCtrl'
 import importMarkdown from '../utils/importMarkdown'
-import { getImageInfo } from '../utils/checkEditImage'
 import Cursor from '../selection/cursor'
 
 const prototypes = [
@@ -170,19 +169,7 @@ class ContentState {
   }
 
   postRender () {
-    // Show image selector when create a inline image by menu/shortcut/or just input `![]()`
-    const startNode = selection.getSelectionStart()
-    if (startNode) {
-      const imageWrapper = startNode.closest('.ag-inline-image')
-      if (imageWrapper && imageWrapper.classList.contains('ag-empty-image')) {
-        const imageInfo = getImageInfo(imageWrapper)
-        this.muya.eventCenter.dispatch('muya-image-selector', {
-          reference: imageWrapper,
-          imageInfo,
-          cb: () => {}
-        })
-      }
-    }
+    // do nothing.
   }
 
   render (isRenderCursor = true) {
